+++ src/renderer/components/Radio.tsx
@@ -9,6 +9,7 @@ type RadioProps = {
   className?: string
   name?: string
   subtitle?: string
+  icon?: HTMLElement
 }
 
 export default function Radio({
@@ -19,6 +20,7 @@ export default function Radio({
   className,
   name,
   subtitle,
+  icon,
 }: RadioProps) {
   const id: string = Math.floor(Math.random() * 10000).toString()
   return (
@@ -31,6 +33,7 @@ export default function Radio({
         value={value}
         defaultChecked={Boolean(selected)}
       />
+      { icon }
       <label htmlFor={id} className={classNames(!subtitle && 'no-subtitle')}>
         <span>{label}</span>
         {subtitle && <span>{subtitle}</span>}