+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - fix clipboard not working in webxdc apps
 - fix `target=_blank` links in html emails don't work #3408
 
+### Added
+
+ - add video chat instances favicons to selection dialog
+
 <a id="1_40_4"></a>
 
 ## [1.40.4] - 2023-09-14